remove abc
